Fixed Removal

+++ remove-application.sh
@@ -1,14 +1,21 @@
+#!/bin/bash
+
 remove=(
-  "1password:$HOME/.local/share/application/1password.desktop"
-  "Figma:$HOME/.local/share/application/Figma.desktop"
-  "Github:$HOME/.local/share/application/Github.desktop"
-  "Dropbox:$HOME/.local/share/application/Dropbox.desktop"
+  "1password:$HOME/.local/share/applications/1password.desktop"
+  "Figma:$HOME/.local/share/applications/Figma.desktop"
+  "Github:$HOME/.local/share/applications/Github.desktop"
+  "Dropbox:$HOME/.local/share/applications/Dropbox.desktop"
 )
 
-echo "Removing Extra Application..."
+echo "Removing Extra Applications..."
 for entry in "${remove[@]}"; do
-  tgt="${entry##*:}"
   src="${entry%%:*}"
-  rm "$tgt"
-  echo "Removed $src"
-donev
+  tgt="${entry##*:}"
+  
+  if [ -f "$tgt" ]; then
+      rm "$tgt"
+      echo "Removed $src"
+  else
+      echo "Skipped $src (not found)"
+  fi
+done
